Introduce MountpointConfig (#1400)
Extract the configuration and the logic to create a new Mountpoint FUSE
session out of the `cli` module.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------
