+++ mountpoint-s3-fs/src/s3/config.rs
@@ -77,7 +77,7 @@ pub struct PartConfig {
 }
 
 impl PartConfig {
-    pub(crate) fn with_read_write_sizes(read_size_bytes: usize, write_size_bytes: usize) -> Self {
+    pub fn with_read_write_sizes(read_size_bytes: usize, write_size_bytes: usize) -> Self {
         Self {
             read_size_bytes,
             write_size_bytes,
